Vote Manager: Dynamically display items based on their plugin dependence.
New special key: 'required'.
Add this key onto the subkey to specify whether this whole sub section should be displayed by checking plugin in the value loaded or not.

New subfix: `#<plugin_name>`
Use this on the end of a value. if the plugin is not loaded this option will not be displayed on the menu.

Removed health restoring as this is not the purpose of this plugin.